Changed getting database and a commented function
The Preset Database Object has its own methods for loading, saving, and
deleting. This I was planning on a possible switch to a custom tables
database, and organizing data correctly before being used.  The other is
the display method used as a direct function. It's commented since I
didn't get to that point yet.