Changes & Additions to set_query
Most of the setting for Post Types, Taxonomies, and Page Parents are
configured. Created a partial/unfinished method to include the current
parent and terms, and the simular design in APLCore::APL_run needs to be
removed. Additions and testing to post_visibility and possibly some
other variable still need to be worked on.