Additions to APL_Query
Cleaned up some of the notes for set_query and added the methods
query_str_consolidate, tax_query_match, and query_str_merge to minimize
querying, but there is a bug creating an additional unnecessary
post_type. Which is cleaned up with consolidating, but still needs to be
fixed. Bug could be in set_query.