Fixed saving multiple callbacks when saving post objects
Problem was fixed with importing, but had overlooked the problem with WP
executing multiple callback functions with the same tag and priority.
NOTE: It was originally suspected to be a pointer/reference issue.

This fixes the upgrade process when multiple custom objects were used
and already added into WP's Action Hook. Which was still needed per
object, so all hooks are removed replaced by 1 hook.